Module Ogg

* Functions for manipulating ogg streams files using libogg. * *

author Samuel Mimram, Romain Beauxis

exception Not_enough_data

exception Bad_data

exception Out_of_sync

exception End_of_stream

exception Internal_error

module Page : sig ... end

* The page struct encapsulates the data for an Ogg page. * * Ogg pages are the fundamental unit of framing and interleave in an ogg * bitstream. They are made up of packet segments of 255 bytes each. There can * be as many as 255 packet segments per page, for a maximum page size of a * little under 64 kB. This is not a practical limitation as the segments can be * joined across page boundaries allowing packets of arbitrary size. In practice * pages are usually around 4 kB.

module Sync : sig ... end

module Stream : sig ... end

module Skeleton : sig ... end